GeminiChatFlask/GeminiChatFlask/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "google-genai>=1.48.0",
    "google-generativeai>=0.8.5",
]
